+++ src/main/java/com/kuit/conet/auth/kakao/KakaoUserProvider.java
@@ -1,6 +1,6 @@
 package com.kuit.conet.auth.kakao;
 
-import com.kuit.conet.utils.JwtParser;
+import com.kuit.conet.utils.auth.JwtParser;
 import com.kuit.conet.utils.auth.PublicKeyGenerator;
 import com.kuit.conet.common.exception.InvalidTokenException;
 import com.kuit.conet.dto.web.response.auth.KakaoPlatformUserResponseDTO;
