+++ src/main/java/com/kuit/conet/utils/auth/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package com.kuit.conet.utils;
+package com.kuit.conet.utils.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.apache.logging.log4j.util.Strings;
